removing renderprops from set password flow
